fix: update permissions

+++ src/main/java/com/minihome/comment/controller/CommentController.java
@@ -23,8 +23,17 @@ public class CommentController {
 	@GetMapping("/home")
 	public String list(Model model, HttpSession session) { 
 		try {
+			// 현재 로그인한 사용자 정보
+            Integer userId = (Integer) session.getAttribute("userId");
+            
+            if (userId == null) {
+                model.addAttribute("error", "로그인이 필요합니다");
+                return "redirect:/login";
+            }
+            
 			List<Comment> comments = service.readAll();	
 			model.addAttribute("comments", comments);
+			model.addAttribute("currentUserId", userId);
 			
 			// 세션에서 사용자 이름 읽기
             String username = (String) session.getAttribute("username");
@@ -41,7 +50,6 @@ public class CommentController {
 						Model model) {
 
 		System.out.println(">>> POST comment: " + nickname + ", " + commentText);
-//		System.out.println("comment >>>" + comment);
 		
 		Integer userId = (Integer) session.getAttribute("userId");
         Comment comment = new Comment(userId, nickname, commentText);
@@ -55,12 +63,33 @@ public class CommentController {
 	}
 	
 	@GetMapping("/delete")
-	public String remove(@RequestParam("id") int commentId) {
+	public String remove(@RequestParam("id") int commentId, HttpSession session, Model model) {
 		try {
-			service.remove(commentId);
-		} catch (SQLException e) {
-			e.printStackTrace();
-		}
+            // 현재 로그인한 사용자 정보
+            Integer userId = (Integer) session.getAttribute("userId");
+            
+            
+            if (userId == null) {
+                model.addAttribute("error", "로그인이 필요합니다");
+                return "redirect:/login";
+            }
+
+            // 댓글 작성자 정보 조회
+            Integer commentAuthorId = service.findCommentAuthor(commentId); 
+            System.out.println(userId + ", " + commentAuthorId + ", " + commentAuthorId.equals(userId));
+
+            if (!commentAuthorId.equals(userId)) {
+                model.addAttribute("error", "삭제 권한이 없습니다");
+                return "redirect:/home";
+            }
+
+            // 댓글 삭제
+            service.remove(commentId);
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+            model.addAttribute("error", "요청을 처리하는 도중 에러가 발생했습니다. 다시 시도해 주세요.");
+        }
 		return "redirect:/home";
 	}
